Got rid of `macrostep-change-group'. Updated some documentation.
I originally thought that a big change group would be a safer way to
restore the buffer to its original contents when exiting, but it turns
out to cause more problems than it solves:

1) if Emacs runs out of memory for undo information it loses the
original buffer state

2) worse, typing C-/ `undo' while macro expansions were visible would
bork the change group and make it impossible to get back to the
original state.

So it seems safer, on balance, just to rely on the overlays to restore
the original text.

Also added an example to the docs to more clearly explain how
expanding sub-forms out of order works.