Use new `macrostep-macro-start` text property
This property should be placed on the first character of each macro form
in the expanded text.  It was previously conflated with saving
the *structure* represented by the expansion in the
`macrostep-expanded-text` property.  This is (currently) useful for the
Emacs Lisp implementation, but may not be for other implementations,
including SLIME, so it is cleaner to separate the two properties.