Extract Elisp macro environments accurately
Instead of manually parsing all enclosing `(cl-)macrolet` forms,
(ab)uses `macroexpand-all` to get the actual macro-expansion environment
at point.  Macro-expansion environments can then be cached as a text
property to be retrieved when expanding subforms.

As a side effect, this should allow the expansion of local macro forms
within local macro-defining forms such as `(with-js`.  The main
disadvantage is that it is slower, especially when it has to move
backwards repeatedly to find an evaluated position from which to grab
the macro-expansion environment.