Avoid positional arguments to define-minor-mode
Back in Emacs-21.1, `define-minor-mode' grew keyword arguments to
replace its old positional arguments.  Starting with Emacs-28.1
a warning will be omitted if positional arguments are still used.

For backward compatibility one keyword argument has to be provided
if a function body is provided (as otherwise the non-keyword value
argument and the body could not be told apart).